Move Context class outside of version specific namespace
